Remove bricks with gluster_volume.
Adds functions to remove and replace bricks when the module is run.
Whenever some bricks are added and removed at the same time, they are
matched up pairwise and `gluster volume replace-brick` is called on
them. Any extra bricks in either list are then added or removed as
appropriate. Replacing bricks forces a rebalance to be done, regardless
of module setting.